lms/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
mysqlclient>=2.1.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.3.0

